finish category route (all crud operations done)

+++ api/src/services/categoryService.ts
@@ -1,3 +1,4 @@
+import { QueryOptions, UpdateQuery, UpdateWithAggregationPipeline } from 'mongoose'
 import { NotFoundError } from '../helpers/apiError'
 import Category, { CategoryDocument } from '../models/Category'
 
@@ -15,8 +16,12 @@ const findById = async (id: string) => {
   return foundOne
 }
 
-const updateOne = async (id: string, update: Partial<CategoryDocument>) => {
-  const foundOne = await Category.findByIdAndUpdate(id, update)
+const updateOne = async ( 
+  id: string, 
+  update?: UpdateWithAggregationPipeline | UpdateQuery<CategoryDocument> | undefined, 
+  options?: QueryOptions | null | undefined
+) => {
+  const foundOne = await Category.findByIdAndUpdate(id, update, options)
   if (foundOne) {
     return foundOne
   } else {
@@ -25,9 +30,9 @@ const updateOne = async (id: string, update: Partial<CategoryDocument>) => {
 }
 
 const deleteOne = async (id: string) => {
-  const foundOne = await Category.findByIdAndDelete(id)
-  if (foundOne) {
-    return foundOne
+  const foundCategory = await Category.findById(id)
+  if (foundCategory) {
+    return await Category.findByIdAndDelete(id)
   } else {
     throw new NotFoundError()
   }
